// frontend/src/generated/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  /** The `ID` scalar type represents a unique MongoDB identifier in collection. MongoDB by default use 12-byte ObjectId value (https://docs.mongodb.com/manual/reference/bson-types/#objectid). But MongoDB also may accepts string or integer as correct values for _id field. */
  MongoID: any;
};

export type Auth = {
  __typename?: 'Auth';
  accessToken?: Maybe<Scalars['String']>;
  refreshToken?: Maybe<Scalars['String']>;
};

export type Comment = {
  __typename?: 'Comment';
  author?: Maybe<Scalars['String']>;
  body?: Maybe<Scalars['String']>;
  date?: Maybe<Scalars['Date']>;
  attachedToId?: Maybe<Scalars['MongoID']>;
  comments: Array<Comment>;
  _id: Scalars['MongoID'];
};
// ...
export type Post = {
  __typename?: 'Post';
  title?: Maybe<Scalars['String']>;
  author?: Maybe<Scalars['String']>;
  body?: Maybe<Scalars['String']>;
  date?: Maybe<Scalars['Date']>;
  comments: Array<Comment>;
  _id: Scalars['MongoID'];
};
// ...
export enum ThreadsEnum {
  Post = 'POST',
  Comment = 'COMMENT'
}